+++ ForestManagement/changelog.md
@@ -1,5 +1,12 @@
 ﻿# Changelog
 
+## ???
+
+- Upd: Test-FMSchemaLdif - writes a warning when testing an object that has nocreation operation assigned, just modify, and is missing in AD.
+- Upd: Register-FMSchemaLdif - supports specifying the order it is applied at and the configuration Context that provided the settings.
+- Upd: Register-FMSchemaLdif - supports specifying objects, which will not trigger an alert if they are missing.
+- Fix: Test-FMSchemaLdif - fails when comparing certain property-types.
+
 ## 1.0.6 (2020-01-27)
 
 - Metadata update